Rename Thermal Pages
From "Thermal Simple and Thermal (Adv) to:
          "Thermal and Thermal Adv"
