+++ src/vario/ui/PageMenuDisplay.h
@@ -18,7 +18,7 @@ class DisplayMenuPage : public SettingsMenuPage {
   void setting_change(Button dir, ButtonState state, uint8_t count);
 
  private:
-  static constexpr char* labels[6] = {"Back",        "Debug",    "Thermal Lite",
+  static constexpr char* labels[6] = {"Back",        "Debug",    "Thermal",
                                       "Thermal ADV", "Navigate", "Contrast"};
 };
